// apps/api/src/app/filters/enums/filter-field.enum.ts
export enum CaseFilterField {
  SUBJECT_NAME = 'subjectName',
  AGE = 'age',
  INCIDENT_TYPE = 'incidentType',
  INCIDENT_DATE = 'incidentDate',
  CASE_SUMMARY = 'caseSummary',
  ADDRESS_911 = 'address911',
  CALL_TIME_911 = 'callTime911',
  STATE = 'state',
  CITY = 'city',
  ZIP = 'zip',
  ARREST_DATE = 'arrestDate',
  STAGE_ID = 'stageId',
  DUE_DATE = 'dueDate',
  CASE_PRIORITY = 'casePriority',
  CASE_SOURCE = 'caseSource',
  PRIORITY = 'priority',
  CREATED_AT = 'createdAt',
  UPDATED_AT = 'updatedAt',
  ASSIGNEE_IDS = 'assigneeIds',
}