// apps/api/src/app/filters/enums/filter-operator.enum.ts
export enum FilterOperator {
  EQUALS = 'equals',
  NOT_EQUALS = 'notEquals',
  CONTAINS = 'contains',
  STARTS_WITH = 'startsWith',
  ENDS_WITH = 'endsWith',
  IN = 'in',
  NOT_IN = 'notIn',
  GREATER_THAN = 'greaterThan',
  GREATER_THAN_OR_EQUAL = 'greaterThanOrEqual',
  LESS_THAN = 'lessThan',
  LESS_THAN_OR_EQUAL = 'lessThanOrEqual',
  BETWEEN = 'between',
  NOT_BETWEEN = 'notBetween',
  IS_EMPTY = 'isEmpty',
  IS_NOT_EMPTY = 'isNotEmpty',
  EXISTS = 'exists',
  NOT_EXISTS = 'notExists',
  CONTAINS_ANY = 'containsAny',
  CONTAINS_ALL = 'containsAll',
  NOT_CONTAINS_ANY = 'notContainsAny',
  NOT_CONTAINS_ALL = 'notContainsAll',
}